Añadiendo código nativo #10 #7

+++ JMR-Android/app/src/main/java/com/example/alejandro/jmr_android/jmr/JMRColor.java
@@ -34,15 +34,15 @@ public class JMRColor {
         this.azul = azul;
     }
 
-    public long getRojo(){
+    public int getRojo(){
         return rojo;
     }
 
-    public long getAzul(){
+    public int getAzul(){
         return azul;
     }
 
-    public long getVerde(){
+    public int getVerde(){
         return verde;
     }
 }
